Removing unnecessary (double).
git-svn-id: svn://orca.physics.unc.edu/OrcaRoot@68 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Decoders/ORAD3511ADCDecoder.cc
@@ -28,7 +28,7 @@ double ORAD3511ADCDecoder::ReferenceDateOf(UInt_t* record)
 { 
   if(!HasReferenceDate(record)) return 0.0;
 
-  double refDate = (double)(*((double*)(record+2)));
+  double refDate = (*((double*)(record+2)));
   return refDate;
 }
